+++ lib/notion.js
@@ -1,3 +1,3 @@
 export { getAllPosts } from './notion/getAllPosts'
-export { getAllTags } from './notion/getAllTags'
+export { getAllTagsFromPosts } from './notion/getAllTagsFromPosts'
 export { getPostBlocks } from './notion/getPostBlocks'
