+++ lib/notion/getAllPosts.js
@@ -3,6 +3,7 @@ import { NotionAPI } from 'notion-client'
 import { idToUuid } from 'notion-utils'
 import getAllPageIds from './getAllPageIds'
 import getPageProperties from './getPageProperties'
+import { filterPublishedPosts } from './filterPublishedPosts'
 
 export async function getAllPosts() {
   let id = BLOG.notionPageId
@@ -39,14 +40,7 @@ export async function getAllPosts() {
       data.push(properties)
     }
     // remove all the the items doesn't meet requirements
-    const posts = data.filter(post => {
-      return (
-        post.title &&
-        post.slug &&
-        post?.status?.[0] === 'Published' &&
-        (post?.type?.[0] === 'Post' || post?.type?.[0] === 'Page')
-      )
-    })
+    const posts = filterPublishedPosts(data)
 
     // Sort by date
     if (BLOG.sortByDate) {
