+++ pages/feed.js
@@ -2,11 +2,9 @@ import { getAllPosts } from '@/lib/notion'
 import { generateRss } from '@/lib/rss'
 export async function getServerSideProps({ res }) {
   res.setHeader('Content-Type', 'text/xml')
-  let posts = await getAllPosts()
-  posts = posts
-    .filter(post => post.status[0] === 'Published' && post.type[0] === 'Post')
-    .slice(0, 10)
-  const xmlFeed = generateRss(posts)
+  const posts = await getAllPosts()
+  const latestPosts = posts.slice(0, 10)
+  const xmlFeed = generateRss(latestPosts)
   res.write(xmlFeed)
   res.end()
   return {
