+++ pages/search.js
@@ -1,15 +1,12 @@
-import { getAllPosts, getAllTags } from '@/lib/notion'
+import { getAllPosts, getAllTagsFromPosts } from '@/lib/notion'
 import SearchLayout from '@/layouts/search'
 
 export default function search({ tags, posts }) {
   return <SearchLayout tags={tags} posts={posts} />
 }
 export async function getStaticProps() {
-  let posts = await getAllPosts()
-  posts = posts.filter(
-    post => post.status[0] === 'Published' && post.type[0] === 'Post'
-  )
-  const tags = await getAllTags()
+  const posts = await getAllPosts()
+  const tags = getAllTagsFromPosts(posts)
   return {
     props: {
       tags,
